fix(cli): deprecate the use of explicit `workspace_url`, and use `workspace_id` instead of `workspace_name` (https://github.com/leptonai/lepton/issues/1770)
This is to be consistent with our naming convention on the backend.
Checked with @vthinkxie .

Summary:
- For the CLI, now we do not support the use of `--workspace-url`, and
only use `--workspace-id`.
- The old `--workspace-name` is deprecated because we want the frontend
(CLI and web) to be consistent. Later the "workspace name" will be used
to represent the display name on the website.
- Right now, we rely on @vthinkxie 's `portal.lepton.ai` to obtain the
workspace url given an id name.
- For the backend (e.g. mothership), note that we are still using
`workspace name` to denote the workspace id. We don't intend to change
this any time soon. As it is not exposed to the users, it is probably
fine.
- The CLI exposes a hidden option, `--test-only-workspace-url`, that
allows end to end test to pass in the explicit workspace url for now.
You should not use it in actual product use cases.