feat(sdk): clean up the workspace api abstractions (https://github.com/leptonai/lepton/issues/2892)
Instead of having a bunch of individual functions that loads and updates
things, now I put everything into one single place
`WorkspaceInfoLocalRecord` and deals with all queries and updates.

A few differences:
- Now all the management of the local file recording the workspace infos
is managed by the class.
- There are a few other utility functions like getting urls from ids,
etc, and probably we want to move them to utils, but it is fine right
now.
- We always maintain a connection for the current workspace.