feat(sdk): add timeout option to photon run (https://github.com/leptonai/lepton/issues/2997)
including a test to run timeout evaluation.

This is only a SDK side implementation to go together with https://github.com/leptonai/lepton/issues/2839 .
@xiang90 has an experimental workspace and I tested that things worked
end to end (with the only exception that the timeout seemed to have been
explicitly set at 60 second, probably due to earlier experimentation
code in the workspace deployment).

Overall, this seems to be a pretty good starting point for us to get
decent fine-grained control of deployments.